pyproject.toml
[build-system]
requires = ["hatchling>=1.21"]
build-backend = "hatchling.build"

[project]
name = "prumo-assist"
description = "Knowledge, bibliography & academic writing assistant for clinical research. Lives between Zotero, Obsidian, and your agent-host."
readme = "README.md"
license = { file = "LICENSE" }
authors = [
  { name = "Raphael Federicci", email = "rfh.federicci@gmail.com" },
]
requires-python = ">=3.11"
keywords = [
  "research",
  "knowledge-management",
  "bibliography",
  "zotero",
  "obsidian",
  "academic-writing",
  "clinical-research",
  "claude-code-plugin",
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Scientific/Engineering",
  "Topic :: Text Processing :: Markup :: Markdown",
]
dependencies = [
  "typer>=0.12",
  "rich>=13.7",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "pyyaml>=6.0",
  "jinja2>=3.1",
]
dynamic = ["version"]

[project.urls]
Homepage = "https://github.com/raphaelfh/prumo-assist"
Repository = "https://github.com/raphaelfh/prumo-assist"
Issues = "https://github.com/raphaelfh/prumo-assist/issues"

[project.scripts]
prumo = "prumo_assist.cli:app"

[project.optional-dependencies]
dev = [
  "pytest>=8.1",
  "pytest-cov>=5.0",
  "ruff>=0.4",
  "mypy>=1.10",
  "types-PyYAML>=6.0",
]

[tool.hatch.version]
path = "src/prumo_assist/_version.py"

[tool.hatch.build.targets.wheel]
packages = ["src/prumo_assist"]

[tool.hatch.build.targets.wheel.force-include]
"templates" = "prumo_assist/_templates"

[tool.ruff]
line-length = 100
target-version = "py311"
# Arquivos legados (predam o PR0 da CLI). Limpeza vai junto com a migração
# pra `domains/` nos PR1+.
extend-exclude = [
  ".github/scripts/validate_manifests.py",
  "skills/tabular-eda/scripts/create_eda_notebook.py",
]

[tool.ruff.lint]
select = [
  "E",    # pycodestyle errors
  "W",    # pycodestyle warnings
  "F",    # pyflakes
  "I",    # isort
  "B",    # flake8-bugbear
  "UP",   # pyupgrade
  "SIM",  # flake8-simplify
  "RUF",  # ruff-specific
]
# E501: line length handled by formatter
# RUF001/RUF002/RUF003: caracteres unicode "ambíguos" — projeto é em português
ignore = ["E501", "RUF001", "RUF002", "RUF003"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["B011"]
# B008/B006: Typer.Option/Annotated requer chamadas no default — padrão da lib
"src/prumo_assist/cli.py" = ["B006", "B008"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unreachable = true
files = ["src/prumo_assist", "tests"]

[[tool.mypy.overrides]]
module = ["yaml.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
pythonpath = ["src"]
